PiperTextToSpeech/requirements.txt
piper-tts
onnxruntime-gpu
sounddevice
numpy